% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{add_variables}
\alias{add_variables}
\title{Extract variables and add}
\usage{
add_variables(dat, phases, vars, keep_vars = NULL)
}
\arguments{
\item{dat}{table with data, including raw pupil/x/y data}

\item{phases}{table with phase messages}

\item{vars}{table with var messages}

\item{keep_vars}{column names to keep from vars}
}
\description{
Extract variables and add
}
