% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_functions.R
\name{baseline}
\alias{baseline}
\title{baseline}
\usage{
baseline(dat, baselineRange = c(-200, 0), by = NULL)
}
\arguments{
\item{dat}{pupil table object}

\item{baselineRange}{timepoints to baseline on}

\item{by}{by}
}
\value{
A data.table with samples
}
\description{
baseline
}
