% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{blink_report}
\alias{blink_report}
\title{Export a blink report}
\usage{
blink_report(
  dat,
  folder = "blink_reports",
  trials_per_page = 20,
  period = c(-Inf, Inf)
)
}
\arguments{
\item{dat}{table with pupil data, including raw pupil}
}
\description{
Export a blink report
}
