% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_functions.R
\name{blinkreconstruct}
\alias{blinkreconstruct}
\title{detect blinks}
\usage{
blinkreconstruct(
  trace,
  vt = 15,
  maxdur = 500,
  margin = 10,
  smooth_winlength = 21,
  std_thr = 3
)
}
\arguments{
\item{vt}{A pupil velocity threshold. Lower thresholds more easily trigger blinks.}

\item{maxdur}{The maximum duration (in samples) for a blink. Longer blinks are not reconstructed.}

\item{margin}{The margin to take around missing data}

\item{smooth_winlength}{The width of the smoothing window. This should be an odd integer.}

\item{std_thr}{std_thr}

\item{dat}{pupil table object}
}
\value{
A data.table with samples
}
\description{
detect blinks
}
