% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_functions.R
\name{detect_blinks}
\alias{detect_blinks}
\title{detect blinks}
\usage{
detect_blinks(dat, maxDeltaDilation = 16)
}
\arguments{
\item{dat}{pupil table object}

\item{maxDeltaDilation}{dilation threshold}
}
\value{
A data.table with samples
}
\description{
detect blinks
}
