% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_functions.R
\name{downsample}
\alias{downsample}
\title{Downsample}
\usage{
downsample(data, samptime, By = "trial")
}
\arguments{
\item{dat}{pupil table object}

\item{by}{by}

\item{Hz}{downsample to}
}
\value{
A data.table with samples
}
\description{
Downsample
}
