% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_functions.R
\name{expand_blinks}
\alias{expand_blinks}
\title{Expand Blinks}
\usage{
expand_blinks(dat, rejectionWindow = 50)
}
\arguments{
\item{dat}{pupil table object}

\item{rejectionWindow}{ms padding}
}
\value{
A data.table with samples
}
\description{
Expand Blinks
}
