% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_functions.R
\name{getBlinkList}
\alias{getBlinkList}
\title{return list of blinks}
\usage{
getBlinkList(strace, blinkparams)
}
\arguments{
\item{strace}{pupil trace}

\item{by}{by}

\item{win}{number of samples to smooth over}
}
\value{
pupil table object with column with smoothed signal
}
\description{
return list of blinks
}
