% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_functions.R
\name{getMicSaccList}
\alias{getMicSaccList}
\title{get Microsaccccade list}
\usage{
getMicSaccList(xy_dat, saccparams)
}
\arguments{
\item{xy_dat}{xy coordinates of gaze}
}
\value{
A data.table with samples
}
\description{
get Microsaccccade list
}
