% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_functions.R
\name{interpolateblinks}
\alias{interpolateblinks}
\title{interpolate blinks}
\usage{
interpolateblinks(dat, type = "linear", maxgap = 500)
}
\arguments{
\item{dat}{pupil table object}

\item{type}{type of interpolation}
}
\value{
A data.table with samples
}
\description{
interpolate blinks
}
