% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{parse_asc_file}
\alias{parse_asc_file}
\title{Load a Matrix}
\usage{
parse_asc_file(infile)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A data.table with samples
}
\description{
This functions load an ASC file
}
