% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_functions.R
\name{timelock}
\alias{timelock}
\title{Timelock trace}
\usage{
timelock(data, phase_name)
}
\arguments{
\item{phase_name}{of which the start will be new 0 point}

\item{dat}{pupil table object}

\item{by}{by}
}
\value{
A data.table with samples
}
\description{
Timelock trace
}
