% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_functions.R
\name{TP_eyelink}
\alias{TP_eyelink}
\title{Extracts blinks from eyelink data}
\usage{
TP_eyelink(data, blinkdat, margin = 100)
}
\arguments{
\item{margin}{The margin to take around missing data}

\item{dat}{pupil table object}
}
\value{
A data.table with samples
}
\description{
Extracts blinks from eyelink data
}
