% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_functions.R
\name{TP_velocity}
\alias{TP_velocity}
\title{Blink detection using method described by Mathot}
\usage{
TP_velocity(data, vt = 5, maxdur = 500, margin = 100, smooth_winlength = 21)
}
\arguments{
\item{data}{pupil table object}

\item{vt}{A pupil velocity threshold. Lower thresholds more easily trigger blinks.}

\item{maxdur}{The maximum duration (in samples) for a blink. Longer blinks are not reconstructed.}

\item{margin}{The margin to take around missing data}

\item{smooth_winlength}{The width of the smoothing window. This should be an odd integer.}

\item{std_thr}{std_thr}
}
\description{
Blink detection using method described by Mathot
}
