% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{trace_reports}
\alias{trace_reports}
\title{Export a trace processing report}
\usage{
trace_reports(
  dat,
  folder = "trace_reports",
  tag = "traces",
  trials_per_page = 20,
  period = c(-Inf, Inf)
)
}
\arguments{
\item{dat}{table with data, including raw pupil/x/y data}

\item{folder}{where to save reports}

\item{tag}{add this tag to filenames}

\item{trials_per_page}{trials per page}

\item{period}{start and end time to plot}
}
\description{
Export a trace processing report
}
