% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_functions.R
\name{traceprocessor}
\alias{traceprocessor}
\title{trace processor}
\usage{
traceprocessor(data, blinkdat = NULL)
}
\arguments{
\item{data}{eyetracker table object with raw samples}

\item{blinkdat}{blink information}
}
\value{
tracker table with processed traces
}
\description{
trace processor
}
