% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_functions.R
\name{downsample}
\alias{downsample}
\title{Downsample}
\usage{
downsample(data, samptime, By = "trial", ds_cols = NULL)
}
\arguments{
\item{samptime}{New sample duration}

\item{ds_cols}{Columns with data that require downsampling}

\item{dat}{pupil table object}

\item{by}{by}
}
\value{
A data.table with samples
}
\description{
Downsample
}
