% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{parse_asc_file}
\alias{parse_asc_file}
\title{This functions load an ASC file}
\usage{
parse_asc_file(
  infile,
  keep_vars = NULL,
  timelock_to = NULL,
  samptime = NULL,
  baseline_period = c(NULL, NULL),
  stime = NULL,
  etime = NULL,
  blinkparams = NULL,
  saccparams = NULL
)
}
\arguments{
\item{infile}{Path to the input file}

\item{keep_vars}{column names to keep from vars that are passed as messages to the eyelink, allows you to specify which variables from the OS-experiment will be retained.}

\item{timelock_to}{if passed, time variable is set to 0 at onset of a phase}

\item{samptime}{if passed, data is downsampled to this sample time (ms)}

\item{baseline_period}{pupil trace baseline}

\item{TP}{Specify which type of trace processor}
}
\value{
A data.table with samples
}
\description{
This functions load an ASC file
}
