% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_functions.R
\name{preprocess}
\alias{preprocess}
\title{load and preprocess data from an asc file}
\usage{
preprocess(file, keep_raw = T, timelock_to = NULL, keep_vars = NULL, vt = 15)
}
\arguments{
\item{timelock_to}{string value of phase on which to timelock to}

\item{dat}{pupil table object}

\item{accept}{leave just one pupil trace instead of keeping intermediate traces}
}
\value{
A data.table with samples
}
\description{
load and preprocess data from an asc file
}
