% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_functions.R
\name{smooth_trace}
\alias{smooth_trace}
\title{smooth a trace}
\usage{
smooth_trace(dat, by = NULL, win = 10)
}
\arguments{
\item{dat}{pupil table object}

\item{by}{by}

\item{win}{number of samples to smooth over}
}
\value{
pupil table object with column with smoothed signal
}
\description{
smooth a trace
}
