% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_functions.R
\name{traceprocessor}
\alias{traceprocessor}
\title{trace processor}
\usage{
traceprocessor(data, blinkparams = NULL, saccparams = NULL)
}
\arguments{
\item{data}{eyetracker table object with raw samples}

\item{blinkdat}{blink information}
}
\value{
Table with processed traces and list of blinks per trial
}
\description{
trace processor
}
